Update index.md
I moved this up, because this affects everyone using the CLI, whereas the Cygwin and Proxy ones are much smaller impact. It would also be good to include the version of the CLI where this started. because it isn't true on the older version of the installer. 

I also made this change to the CF CLI page (https://console.bluemix.net/docs/cli/reference/cfcommands/index.html#cf), which is probably where it will be most relevant.